homework12/pyproject.toml
[project]
name = "MLOps_course_AGH_lab_12"
version = "1.0.0"
description = ""
readme = "README.md"

requires-python = "==3.11.*"

dependencies = [
    "apache-airflow[celery,postgres,standard,uv]==3.1.5",
    "apache-airflow-providers-amazon[s3fs]",
    "s3fs",
    "twelvedata",
    "polars>=1.37.1",
]

[dependency-groups]
dev = [
    "ruff"
]

airflow_common = [
    "apache-airflow[celery,postgres,standard,uv]==3.1.5",
    "requests",
]
